\name{wc14goals-package}
\alias{wc14goals-package}
\alias{wc14goals}
\docType{package}
\title{
~~ FIFA World Cup Finals, Brazil 2014 Goal Data ~~
}
\description{
More about what it does (maybe more than one line)
~~ 2014 FIFA World Cup Final data extracted from the openfootball database originally compiled by Gerald Bauer, Jose Borja, Tarek Amr, et al. ~~
}
\details{
\tabular{ll}{
Package: \tab wc14goals\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-24\cr
License: \tab No license\cr
}
~~ Data only. ~~
}
\author{
Michael O'Dell

Maintainer:
~~ Michael O'Dell <mp_odell@yahoo.com> ~~
}
\references{
~~ openfootball (https://github.com/openfootball) and openmundi (https://github.com/openmundi) ~~
}

\keyword{ package }
\seealso{

}
\examples{

}
